replace u8 with u32 prevent overflow bugs

+++ contracts/src/actions.cairo
@@ -6,51 +6,51 @@ use dojo_arena::models::io::{
 use dojo_arena::models::{ArenaCharacter};
 
 
-const HP_MULTIPLIER: u8 = 10;
-const BASE_HP: u8 = 90;
+const HP_MULTIPLIER: u32 = 10;
+const BASE_HP: u32 = 90;
 
-const ENERGY_MULTIPLIER: u8 = 3;
-const BASE_ENERGY: u8 = 20;
+const ENERGY_MULTIPLIER: u32 = 3;
+const BASE_ENERGY: u32 = 20;
 
 // const NUMBER_OF_PLAYERS: u32 = 4;
 
 const COUNTER_ID: u32 = 99999999;
 
-const FIRST_POS: u8 = 6;
-const SECOND_POS: u8 = 9;
-const RANGE_POS: u8 = 15;
-const MAX_TURNS: u8 = 25;
+const FIRST_POS: u32 = 6;
+const SECOND_POS: u32 = 9;
+const RANGE_POS: u32 = 15;
+const MAX_TURNS: u32 = 25;
 
-const AGI_INITIATIVE_MODIFIER: u8 = 25;
+const AGI_INITIATIVE_MODIFIER: u32 = 25;
 // Base_Quick_Atack_Initiative
-const QUICK_ATC_INI: u8 = 10;
+const QUICK_ATC_INI: u32 = 10;
 // Base_Precise_Atack_Initiative
-const PRECISE_ATC_INI: u8 = 15;
+const PRECISE_ATC_INI: u32 = 15;
 // Base_Heavy_Atack_Initiative
-const HEAVY_ATC_INI: u8 = 20;
+const HEAVY_ATC_INI: u32 = 20;
 // Base Move Initiative
-const MOVE_INI: u8 = 8;
-const REST_INI: u8 = 8;
+const MOVE_INI: u32 = 8;
+const REST_INI: u32 = 8;
 
-const QUICK_ATC_ENERGY: u8 = 2;
-const PRECISE_ATC_ENERGY: u8 = 4;
-const HEAVY_ATC_ENERGY: u8 = 6;
+const QUICK_ATC_ENERGY: u32 = 2;
+const PRECISE_ATC_ENERGY: u32 = 4;
+const HEAVY_ATC_ENERGY: u32 = 6;
 
-const QUICK_ATC_DAMAGE: u8 = 5;
-const PRECISE_ATC_DAMAGE: u8 = 10;
-const HEAVY_ATC_DAMAGE: u8 = 20;
+const QUICK_ATC_DAMAGE: u32 = 5;
+const PRECISE_ATC_DAMAGE: u32 = 10;
+const HEAVY_ATC_DAMAGE: u32 = 20;
 
 const QUICK_HIT_CHANCE: u128 = 80;
 const PRECISE_HIT_CHANCE: u128 = 60;
 const HEAVY_HIT_CHANCE: u128 = 35;
 
-const REST_RECOVERY: u8 = 5;
+const REST_RECOVERY: u32 = 5;
 
-const MAX_LEVEL: u8 = 9;
-const MAX_STRENGTH: u8 = 9;
-const MAX_AGILITY: u8 = 9;
-const MAX_VITALITY: u8 = 9;
-const MAX_STAMINA: u8 = 9;
+const MAX_LEVEL: u32 = 9;
+const MAX_STRENGTH: u32 = 9;
+const MAX_AGILITY: u32 = 9;
+const MAX_VITALITY: u32 = 9;
+const MAX_STAMINA: u32 = 9;
 
 #[starknet::interface]
 trait IActions<TContractState> {
@@ -62,7 +62,9 @@ trait IActions<TContractState> {
     fn register(self: @TContractState, arena_id: u32);
     fn play(self: @TContractState, arena_id: u32);
     fn level_up(self: @TContractState);
-    fn assign_points(self: @TContractState, strength: u8, agility: u8, vitality: u8, stamina: u8);
+    fn assign_points(
+        self: @TContractState, strength: u32, agility: u32, vitality: u32, stamina: u32
+    );
     fn update_strategy(self: @TContractState, strategy: ClassHash);
     fn battle(self: @TContractState, c1: ArenaCharacter, c2: ArenaCharacter) -> ArenaCharacter;
 }
@@ -319,7 +321,7 @@ mod actions {
         }
 
         fn assign_points(
-            self: @ContractState, strength: u8, agility: u8, vitality: u8, stamina: u8
+            self: @ContractState, strength: u32, agility: u32, vitality: u32, stamina: u32
         ) {
             let world = self.world_dispatcher.read();
             let player = get_caller_address();
@@ -373,9 +375,9 @@ mod actions {
                 hp: c2.hp, position: SECOND_POS, energy: c2.energy, consecutive_rest_count: 0,
             };
 
-            let mut c1_initiative: u8 = 0;
-            let mut c2_initiative: u8 = 0;
-            let mut turns: u8 = 0;
+            let mut c1_initiative: u32 = 0;
+            let mut c2_initiative: u32 = 0;
+            let mut turns: u32 = 0;
 
             let mut winner = c1.clone();
             loop {
